StorybookでTemplate.bind()の型がanyになってしまう不具合の解消
