+++ package.json
@@ -7,7 +7,7 @@
   ],
   "license": "MIT",
   "scripts": {
-    "build": "tsc",
+    "build": "tsc -p tsconfig-build.json",
     "lint": "eslint src/**",
     "lint:fix": "eslint --fix src/**",
     "format": "prettier --write src/**",
