+++ tsconfig-build.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig",
+  "exclude": ["src/**/*.stories.tsx"]
+}
